chore: rename some components
[#179355268](https://www.pivotaltracker.com/story/show/179355268)
