+++ internal/encryption/passwordparser/parser.go
@@ -1,4 +1,4 @@
-package parser
+package passwordparser
 
 import (
 	"encoding/json"
